requirements.txt
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
esy-osm-pbf==0.0.1
esy-osmfilter==1.0.7
future==0.18.2
geocoder==1.38.1
idna==2.9
kiwisolver==1.2.0
matplotlib==3.2.1
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy==1.18.4
opencv-python==4.2.0.34
pandas==1.0.4
Pillow==7.1.2
protobuf==3.11.3
pygame==1.9.6
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
ratelim==0.1.6
requests==2.23.0
scipy==1.4.1
seaborn==0.10.1
sip==4.19.13
six==1.15.0
tornado==6.0.4
urllib3==1.25.9
wincertstore==0.2
